fix: transitive headers -> direct headers; set strict compilation

+++ task1.3/main.cpp
@@ -4,6 +4,8 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
+#include <limits>
+#include <ranges>
 
 #include "../common/edge.hpp"
 #include "../common/graph.hpp"
@@ -90,7 +92,7 @@ class Dinic {
         }
 
         const auto arcs = make_arcs_from(u, graph);
-        while (next_arc_index[u] < arcs.size()) {
+        while (next_arc_index[u] < (i64) arcs.size()) {
             const auto &e = arcs[next_arc_index[u]];
             const auto &v = e.to();
